nolej-models/requirements.txt
joblib==1.0.0
nltk==3.5
numpy==1.19.5
protobuf==3.14.0
psutil==5.8.0
pyparsing==2.4.7
requests==2.25.1
sentencepiece==0.1.95
tokenizers==0.10.3
torch==1.7.1
tqdm==4.56.0
transformers==4.8.2
urllib3==1.26.2
pysbd==0.3.4
sentence-transformers==1.2.1
spacy
textrazor
flashtext
python-dotenv
boto3
sagemaker
keybert==0.5.1
